refactor `packages/docusaurus/src/server/types`
