+++ packages/docusaurus-types/src/index.ts
@@ -1,10 +1,10 @@
-import {Loader} from 'webpack';
+import {Loader, Configuration} from 'webpack';
 import {ParsedUrlQueryInput} from 'querystring';
 
 export interface DocusaurusConfig {
   baseUrl: string;
-  favicon?: string;
-  tagline?: string;
+  favicon: string;
+  tagline: string;
   title: string;
   url: string;
   organizationName?: string;
@@ -22,12 +22,15 @@ export interface DocusaurusConfig {
 }
 
 export interface DocusaurusContext {
-  siteConfig?: DocusaurusConfig;
+  siteConfig?: Partial<DocusaurusConfig>;
 }
 
-export type PluginConfig = [string, Object | undefined] | string;
+export interface Preset {
+  plugins?: PluginConfig[];
+  themes?: PluginConfig[];
+}
 
-export type PresetConfig = [string, Object | undefined] | string;
+export type PresetConfig = [string, Object] | [string] | string;
 
 export interface CLIOptions {
   [option: string]: any;
@@ -35,16 +38,46 @@ export interface CLIOptions {
 
 export interface LoadContext {
   siteDir: string;
-  generatedFilesDir?: string;
-  siteConfig: DocusaurusConfig;
-  cliOptions?: CLIOptions;
-  outDir?: string;
-  baseUrl?: string;
+  generatedFilesDir: string;
+  siteConfig: Partial<DocusaurusConfig>;
+  cliOptions: CLIOptions;
+  outDir: string;
+  baseUrl: string;
+}
+
+export interface Props extends LoadContext {
+  routesPaths: string[];
+  plugins: Plugin<any>[];
 }
 
 export interface PluginContentLoadedActions {
   addRoute(config: RouteConfig): void;
   createData(name: string, data: Object): Promise<string>;
+}
+
+export interface Plugin<T> {
+  name: string;
+  loadContent?(): T;
+  contentLoaded?({
+    content,
+    actions,
+  }: {
+    content: T;
+    actions: PluginContentLoadedActions;
+  }): void;
+  postBuild?(props: Props): void;
+  postStart?(props: Props): void;
+  configureWebpack?(config: Configuration, isServer: boolean): Configuration;
+  getThemePath?(): string;
+  getPathsToWatch?(): string[];
+  getClientModules?(): string[];
+}
+
+export type PluginConfig = [string, Object | undefined] | string;
+
+export interface ChunkRegistry {
+  importStatement: string;
+  modulePath: string;
 }
 
 export type Module =
@@ -65,6 +98,10 @@ export interface RouteConfig {
   modules?: RouteModule;
   routes?: RouteConfig[];
   exact?: boolean;
+}
+
+export interface ThemeAlias {
+  [alias: string]: string;
 }
 
 export interface ConfigureWebpackUtils {
