+++ packages/docusaurus/src/client/exports/useDocusaurusContext.ts
@@ -7,7 +7,7 @@
 
 import {useContext} from 'react';
 import context from './context';
-import {DocusaurusContext} from '../types';
+import {DocusaurusContext} from '@docusaurus/types';
 
 function useDocusaurusContext(): DocusaurusContext {
   return useContext(context);
