--- packages/docusaurus/src/client/types.ts
@@ -1,5 +1,0 @@
-import {DocusaurusConfig} from '../server/types';
-
-export interface DocusaurusContext {
-  siteConfig?: DocusaurusConfig;
-}
