+++ packages/docusaurus/src/server/loadSetup.ts
@@ -7,7 +7,7 @@
 
 import path from 'path';
 import {load} from './index';
-import {Props} from './types';
+import {Props} from '@docusaurus/types';
 
 // Helper methods to setup dummy/fake projects
 export const loadSetup = async (name: string): Promise<Props> => {
