+++ packages/docusaurus/src/server/routes.ts
@@ -8,7 +8,7 @@
 import {genChunkName} from '@docusaurus/utils';
 import _ from 'lodash';
 import {stringify} from 'querystring';
-import {ChunkRegistry, Module, RouteConfig, RouteModule} from './types';
+import {ChunkRegistry, Module, RouteConfig, RouteModule} from '@docusaurus/types';
 
 function getModulePath(target: Module): string {
   if (typeof target === 'string') {
