+++ packages/docusaurus/src/webpack/server.ts
@@ -11,7 +11,7 @@ import {Configuration} from 'webpack';
 import merge from 'webpack-merge';
 import WebpackNiceLog from 'webpack-nicelog';
 
-import {Props} from '../server/types';
+import {Props} from '@docusaurus/types';
 import {createBaseConfig} from './base';
 import WaitPlugin from './plugins/WaitPlugin';
